feat: development service, refactor service names
@alessandrotenaglia
Added a development service which works on the latex-env:dev image, where the internal 'latex' user has sudo privileges and is password-protected.
Such an image/environment is intended for development tasks, e.g., finding out which missing package should be installed and then added to latex-env:latest.
In case one needs to set up the development environment while working on a project, all that needs to be done is modifying .devcontainer/devcontainer.json:8-10, changing the service name to 'latex-dev'.
Note that in latex-env:latest, the internal user has no privileges whatsoever.
Service names had to be refactored to support the new change.